refactor: split tab-tracking.ts into UI state + API/storage (coherence 0.40)
Extract Chrome API calls and storage operations (handleStopTracking,
handleUrlClick, handleTrackPageClick) into tab-tracking-api.ts. UI state
rendering functions stay in tab-tracking.ts.
